Remove horizontal spacer
This seems to have been causing segmentation faults when closing the
main window if the rotation dialog had been opened. I assume this was
because I was using the same spacer object in two places. The second use
didn't affect anything anyway, because it was right next to the first
one