Fix `GlobalSettings` singleton behaviour
Previously, `GlobalSettings.__init__()` would be called when lintrans
was first loaded because it was needed for `global_settings` in
`lintrans.global_settings`.

Since `GlobalSettings.__init__()` needs to run `sys.executable` to see
if it's lintrans, this would create an infinite cascade of processes
calling themselves before they could print their versions. I
accidentally OOM-killed my whole DE with this when I compiled lintrans.

This commit means that the `GlobalSettings` object is now created
lazily. `__init__()` is called once when it's first needed, and then
never again.